Rename bibtex default classes to avoid confusion
