+++ src/nl/hannahsten/texifyidea/completion/BibtexTypeTokenProvider.kt
@@ -7,7 +7,7 @@ import com.intellij.codeInsight.lookup.LookupElementBuilder
 import com.intellij.util.PlatformIcons
 import com.intellij.util.ProcessingContext
 import nl.hannahsten.texifyidea.completion.handlers.TokenTypeInsertHandler
-import nl.hannahsten.texifyidea.lang.BibtexDefaultEntry
+import nl.hannahsten.texifyidea.lang.BibtexDefaultEntryType
 import nl.hannahsten.texifyidea.lang.LatexPackage
 import java.util.*
 
@@ -18,7 +18,7 @@ object BibtexTypeTokenProvider : CompletionProvider<CompletionParameters>() {
 
     override fun addCompletions(parameters: CompletionParameters, context: ProcessingContext, result: CompletionResultSet) {
         result.addAllElements(
-            BibtexDefaultEntry.values().map {
+            BibtexDefaultEntryType.values().map {
                 LookupElementBuilder.create(it, it.token)
                     .withPresentableText(it.token)
                     .bold()
@@ -29,11 +29,11 @@ object BibtexTypeTokenProvider : CompletionProvider<CompletionParameters>() {
         )
     }
 
-    private fun tags(entry: BibtexDefaultEntry): String {
+    private fun tags(entry: BibtexDefaultEntryType): String {
         return " {" + entry.required.joinToString { it.toString().lowercase(Locale.getDefault()) } + "}"
     }
 
-    private fun packageName(entry: BibtexDefaultEntry): String {
+    private fun packageName(entry: BibtexDefaultEntryType): String {
         return when (val dependency = entry.dependency) {
             LatexPackage.DEFAULT -> ""
             else -> " (${dependency.name})"
